[no ci] WIP new module simulation to avoid circular import
